+++ crates/util/src/test_util.rs
@@ -162,7 +162,7 @@ pub fn assert_result<TA, TE, E>(
 
 /// Enable tracing output. Call this in a test to enable logging
 #[deprecated(note = "Debugging only; remove when done")]
-pub fn initialize_tracing() {
+pub fn initialize_tracing(level: LevelFilter) {
     let subscriber = tracing_subscriber::fmt::layer()
         .with_writer(io::stderr)
         .with_target(false)
@@ -170,7 +170,7 @@ pub fn initialize_tracing() {
         .without_time()
         .with_filter(
             Targets::new()
-                .with_target("slumber", LevelFilter::TRACE)
+                .with_target("slumber", level)
                 .with_default(LevelFilter::WARN),
         );
     tracing_subscriber::registry().with(subscriber).init();